Redirect to HTTPS
It appears that without redirecting to HTTPS, even with using
"https://" in the redirect for the FB OAuth dialog that we are
redirected to an "http" URL when we visit the site originally not in
HTTPS, which causes transmission of secrets in the clear.  Do the
right thing here and just redirect to HTTPS to begin with.  Perhaps a
little better would be Flask middleware, but we don't need it with the
number of URIs that we have.